Remove MD calcs and move conformer search to Calculation interface.
